requirements.txt
dj-database-url==0.4.1
Django==1.11.7
django-argonauts==1.2.0
django-celery-results==1.2.1
django-debug-toolbar==1.11
django-google-auth2==0.0.8
django-import-export==1.2.0
django-mfa==2.0
django-modelcluster==5.0.1
django-otp==0.7.5
django-ranged-response==0.2.0
django-redis==4.11.0
django-redis-cache==2.1.0
django-simple-captcha==0.5.12
django-simpleui==3.9
django-taggit==1.2.0
django-tinymce==2.8.0
django-treebeard==4.3.1
djangorestframework==3.11.0
dnspython==1.16.0
dominate==2.5.1
gunicorn==19.6.0
Jinja2==2.11.1
mysqlclient==1.4.6
openpyxl==3.0.3
post==2019.4.13
psycopg2==2.7
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==12.3.0
python-u2flib-server==5.0.0
python-whois==0.7.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
sqlparse==0.3.0
tornado==6.0.3
Unidecode==1.1.1
urllib3==1.25.7
uWSGI==2.0.18
